refactor(filters): simplify filters
Big file was divided in small modules
